requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.10.7
future==0.18.3
gevent==22.10.2
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
promise==2.3
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
text-unidecode==1.3
tldextract==3.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
